initial Github Commit
This is a complete overhaul for this app. Updating the UI, graphics and
support for iOS 8 and new iPhone devices. Work in progress, fixed all
view controllers except the summary. Need to work on animations as well
as tracking throughout the process & add the saving feature to the
summary page (connected with the load feature at the beginning).